requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.27
pydantic>=2.6
httpx>=0.27
python-frontmatter>=1.1.0
chromadb>=0.5.0
sentence-transformers>=3.0.0
pytest>=8.0
jsonschema>=4.21
python-dotenv>=1.0